# openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Future Me API",
    "description": "Future Me is a time capsule letter platform that allows users to send letters to their future self or loved ones, scheduled for delivery on a future date. This API provides endpoints for user authentication, letter creation and management, draft system, scheduled delivery, notifications, and media attachments.",
    "version": "1.0.0",
    "contact": {
      "name": "Future Me Support",
      "email": "support@futureme.com"
    }
  },
  "servers": [
    {
      "url": "https://backend.hyfuture.app/api/v1",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:4001/api/v1",
      "description": "Local Development Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token. Get it from the login or verify-auth-code endpoints."
      }
    },
    "schemas": {
      "RegisterUserRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "example": "John Doe",
            "description": "Optional user name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 100,
            "format": "password",
            "example": "Testpass123!"
          }
        }
      },
      "VerifyAuthCodeRequest": {
        "type": "object",
        "required": ["userId", "code"],
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "example": "ce285529-0508-4112-bd64-56d4156af2d6"
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "example": "654353",
            "description": "6-digit verification code"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "adamma.okoro@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "Testpass123!"
          }
        }
      },
...
      "ForgotPasswordRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "adamma.okoro@example.com"
          }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "required": ["email", "password", "confirmPassword", "code"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "adamma.okoro@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 100,
            "format": "password",
            "example": "Newpass123!"
          },
          "confirmPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 100,
            "format": "password",
            "example": "Newpass123!"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "Password reset code"
          }
        }
      },
      "ResendAuthCodeRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "adamma.okoro@example.com"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "isEmailVerified": {
            "type": "boolean"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "newUser": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "message": {
            "type": "string",
            "example": "User registered successfully. Please verify your email."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        }
      },
...
      "OAuthCallbackResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        }
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["User Onboarding"],
        "summary": "Register User",
        "description": "Create a new Future Me user account. Email must be unique; password requires ≥8 characters. A verification code will be sent to the provided email address. Guest letters sent from this email will be automatically linked to the account.",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully. Verification code sent to email.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input or user already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-auth-code": {
      "post": {
        "tags": ["User Onboarding"],
        "summary": "Verify Auth Code",
        "description": "Complete email verification using the 6-digit code sent to the user's email. Returns access and refresh tokens upon successful verification. Any guest letters sent from this email will be automatically linked to the account.",
        "operationId": "verifyAuthCode",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyAuthCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully. Access and refresh tokens returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid or expired verification code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid verification code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Sessions & Tokens"],
        "summary": "Login User",
        "description": "Authenticate a verified user and receive access/refresh tokens. The account must be verified before login.",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. Returns access and refresh tokens.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials or unverified account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/renew-tokens": {
      "post": {
        "tags": ["Sessions & Tokens"],
        "summary": "Renew Tokens",
        "description": "Use the refreshToken httpOnly cookie to obtain a new access token and rotate the refresh token cookie, without re-authenticating. No request body is required.",
        "operationId": "renewTokens",
        "responses": {
          "200": {
            "description": "Tokens renewed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Sessions & Tokens"],
        "summary": "Logout User",
        "description": "Logs out the user by revoking the current refresh token associated with the refreshToken httpOnly cookie and clearing that cookie.",
        "operationId": "logoutUser",
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing refresh token cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Password Recovery"],
        "summary": "Forgot Password",
        "description": "Trigger a password reset email/code for the user account. A 6-digit reset code will be sent to the registered email address.",
        "operationId": "forgotPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset code sent to email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["Password Recovery"],
        "summary": "Reset Password",
        "description": "Confirm the reset code and set a new password. The code must match the one sent via email.",
        "operationId": "resetPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid code, password mismatch, or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or expired reset code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/resend-verification-code": {
      "post": {
        "tags": ["Code Resend"],
        "summary": "Resend Verification Code",
        "description": "Resends the email verification auth code for unverified users. Use this if you didn't receive the initial verification email.",
        "operationId": "resendVerificationCode",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendAuthCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification code resent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid email or user already verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/resend-password-code": {
      "post": {
        "tags": ["Code Resend"],
        "summary": "Resend Password Code",
        "description": "Triggers another password-reset auth code using the same flow as forgot password. Use this if you didn't receive the password reset code.",
        "operationId": "resendPasswordCode",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset code resent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/google/callback": {
      "get": {
        "tags": ["OAuth"],
        "summary": "Google OAuth Callback",
        "description": "Callback endpoint for Google OAuth. Handles both registration and login flows based on state parameter. Returns access token and user information. Sets refreshToken cookie.",
        "operationId": "googleCallback",
        "responses": {
          "200": {
            "description": "OAuth authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthCallbackResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Email already exists (for registration) or account not found (for login)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - OAuth authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/facebook/callback": {
      "get": {
        "tags": ["OAuth"],
        "summary": "Facebook OAuth Callback",
        "description": "Callback endpoint for Facebook OAuth. Handles both registration and login flows based on state parameter. Returns access token and user information. Sets refreshToken cookie.",
        "operationId": "facebookCallback",
        "responses": {
          "200": {
            "description": "OAuth authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthCallbackResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Email already exists (for registration) or account not found (for login)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - OAuth authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}